Move garbage collector functions to operation_controller.go
