--- cluster-agent/controllers/managed-gitops/operation_garbage_collector.go
@@ -1,80 +1,0 @@
-/*
-Copyright 2021.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package controllers
-
-import (
-	"context"
-	"time"
-
-	"github.com/go-logr/logr"
-	"github.com/redhat-appstudio/managed-gitops/backend-shared/config/db"
-	"sigs.k8s.io/controller-runtime/pkg/log"
-)
-
-const (
-	garbageCollectionInterval = 10 * time.Minute
-)
-
-type garbageCollector struct {
-	db db.DatabaseQueries
-}
-
-// NewGarbageCollector creates a new instance of garbageCollector
-func NewGarbageCollector(dbQueries db.DatabaseQueries) *garbageCollector {
-	return &garbageCollector{
-		db: dbQueries,
-	}
-}
-
-func (g *garbageCollector) StartGarbageCollector() {
-	g.startGarbageCollectionCycle()
-}
-
-func (g *garbageCollector) startGarbageCollectionCycle() {
-	func() {
-		for {
-			// garbage collect the operations after a specified interval
-			<-time.After(garbageCollectionInterval)
-
-			ctx := context.Background()
-			log := log.FromContext(ctx)
-
-			// get the  failed/completed operations with non-zero gc interval
-			operations := []db.Operation{}
-			err := g.db.ListOperationsToBeGarbageCollected(ctx, &operations)
-			if err != nil {
-				log.Error(err, "failed to list operations ready for garbage collection")
-			}
-
-			g.garbageCollectOperations(ctx, operations, log)
-		}
-	}()
-}
-
-func (g *garbageCollector) garbageCollectOperations(ctx context.Context, operations []db.Operation, log logr.Logger) {
-	for _, operation := range operations {
-		// last_state_update + gc_expiration_time < time.Now
-		if operation.Last_state_update.Add(operation.GetGCExpirationTime()).Before(time.Now()) {
-			_, err := g.db.DeleteOperationById(ctx, operation.Operation_id)
-			if err != nil {
-				log.Error(err, "failed to delete operation", operation.Operation_id)
-				continue
-			}
-			log.Info("successfully garbage collected operation", "operation_id", operation.Operation_id)
-		}
-	}
-}
